lox lectures finished. Now the fun begins. Going through all challenges. To build

+++ value.h
@@ -11,7 +11,7 @@ typedef struct ObjString ObjString;
 #ifdef NAN_BOXING 
 
 #define SIGN_BIT    ((uint64_t)0x8000000000000000)
-#define QNAN        ((uint8_t)0x7ffc000000000000)
+#define QNAN        ((uint64_t)0x7ffc000000000000)
 
 #define TAG_NIL   1   // 01
 #define TAG_FALSE 2   // 10
